refactor: 불필요한 console.log 제거 및 DogType 인터페이스 위치 이동
